Use real school images in school page
Refs #42.
