+++ app/components/SchoolMap.js
@@ -1,0 +1,12 @@
+'use strict';
+
+import React from 'react';
+
+class SchoolMap extends React.Component {
+
+  render() {
+    return <div className='school-map'/>;
+  }
+}
+
+export default SchoolMap;
